// carbonFootprintCalculator/src/Storage/StorageValues.Enum.ts
export enum STORAGE_VALUES_MOBILITY {
  NONE = 'NONE',
  CAR = 'CAR',
  PUBLIC_TRANSPORT = 'PUBLIC_TRANSPORT',
  BIKE = 'BIKE',
  FOOT = 'FOOT',
}

export enum STORAGE_VALUES_TRAVEL {
  NONE = 'NONE',
  GERMANY = 'GERMANY',
  EUROPE = 'EUROPE',
  TRANSCONTINENTAL = 'TRANSCONTINENTAL',
}

export enum STORAGE_VALUES_FLIGHT_HOURS {
  NONE = 'NONE',
  LESS_THAN_TWO = 'LESS_THAN_TWO',
  TWO_TO_FOUR = 'TWO_TO_FOUR',
  FOUR_TO_EIGHT = 'FOUR_TO_EIGHT',
  MORE_THAN_EIGHT = 'MORE_THAN_EIGHT',
}

export enum STORAGE_VALUES_FOOD {
  NONE = 'NONE',
  VEGAN = 'VEGAN',
  VEGETARIAN = 'VEGETARIAN',
  MIX_DIET = 'MIX_DIET',
  MEAT_CENTRIC = 'MEAT_CENTRIC',
}

export enum STORAGE_VALUES_LIVING {
  NONE = 'NONE',
  ONLY_ME = 'ONLY_ME',
  TWO_TO_FOUR = 'TWO_TO_FOUR',
  MORE_THAN_FOUR = 'MORE_THAN_FOUR',
}

export enum STORAGE_VALUES_ELECTRICITY {
  NONE = 'NONE',
  MIX = 'MIX',
  GREEN = 'GREEN',
}